add tests for helpers
